Mean vertex calibration (#4638)
* First implementation of MeanVertex calibration

In the direction of the moving average.

Vectors for SMA

Compiling, but I did not check if it works

typos

Processing MeanVertex.

* First bunch of comments by Ruben (a couple left, for next week :-) )

Following review to PR.

following review

clang-format

* Using ConfigurableParams

clang

* remove mTmpMVobjDqTimeEnd, and fixes

* Fix determination of startValidity

clang

* Using Bracket

Using Bracket and changes related to modifications of Utils in dev

clang

+++ Detectors/TOF/calibration/src/TOFChannelCalibrator.cxx
@@ -473,7 +473,7 @@ void TOFChannelCalibrator::finalizeSlot(Slot& slot)
       }
     }
 
-    int fitres = fitGaus(c->getNbins(), histoValues.data(), -(c->getRange()), c->getRange(), fitValues);
+    double fitres = fitGaus(c->getNbins(), histoValues.data(), -(c->getRange()), c->getRange(), fitValues);
 
     if (fitValues[2] < 0)
       fitValues[2] = -fitValues[2];
